scripts/tiendeo_ocr_requirements.txt
requests
beautifulsoup4
Pillow
opencv-python-headless
pytesseract
google-cloud-vision
numpy